add DOTFILES_PATH environment variable check inside daily updater

+++ .bash/functions.sh
@@ -1,6 +1,10 @@
 #!/bin/bash
 
 daily(){
+  if [[ -z "$DOTFILES_PATH" ]]; then
+    echo "DOTFILES_PATH environment variable is not set."
+    return 1
+  fi
   cd ${DOTFILES_PATH};
   git fetch --all --prune && git reset --hard origin/master;
   find . -type f -not -regex "./.git.*" | xargs -L 1 -I {} sh -c "mkdir -p \`dirname ~/{}\`; cp {} ~/{};";
